Speed up similarity by transposing less.
Dot used to transpose it's second argument, which took much longer than
the product it's self. This changes the definition of dot and similarity
so that the second argument is a column matrix, so generally should be
transposed by the user.

sort_by_similarity now transposes the test vector, and uses that to
compare against.