Only store one ID per vector.
If one is not given, it will be automatically generated; index.add now
returns the vector ID.

This does not check for collisions between automatically generated and
user-provided IDs; it is expected that users will either use one or the
other, but it might be worth checking.

Currently, IDs for the in-memory index are generated as integers but may
be any type. IDs for Redis must be strings as they're used as keys.